feat(DeleteStation): 노선에 등록된 역은 삭제할 수 없는 기능 구현 - 노선에 등록된 역을 삭제하려고 하면 에러메시지를 출력하고 역관리 화면으로 돌아감 - fix: 라인 대시보드 이름 잘못 출력되는 문제 수정

+++ src/main/java/subway/dashboard/StationDashboard.java
@@ -94,7 +94,7 @@ public class StationDashboard {
     }
 
     public boolean startOption2(InputView inputView) {
-        if (deleteStation(inputView)) {
+        if (!deleteStation(inputView)) {
             System.out.println();
             return true;
         }
@@ -136,11 +136,16 @@ public class StationDashboard {
 
     public boolean deleteStation(InputView inputView) {
         String submittedStationName = inputView.readDeletingStationName();
+        if (!StationRepository.canDeleteStation(submittedStationName)) {
+            System.out.println(ERROR_CANNOT_DELETE_STATION_ON_LINE);
+            return false;
+        }
+
         if (!StationRepository.deleteStation(submittedStationName)) {
             System.out.println(ERROR_STATION_NAME_NO_EXISTS);
-            return true;
+            return false;
         }
-        return false;
+        return true;
     }
 
     public void showStations() {
